Adding user engagements
Adding user engagements to help admin track how user goes through the database, I also change a bit of the database to make it more concistent.

I also plan to do a lot of changes on visualization of user engagements.

I also plan to improve the change image helper to make it better.