Table history update
I finally finishing up some feature. I added the Omnitags model to ensure some universal functions are covered perfectly, I also add the new history page and archive page. Also trying to minimize some of the functions like loading a page since it tracks a lot of information in the website so I crammed them into a function.

I added the ability to check the history of changes to records that you have made. I ensure that the feature is going well and well written. Adding a lot of buttons is kinda a hassle but it's kinda fun when it's done. I also hardcoded the first field of each history table I just made, which is straight id_history to ensure that I don't overwork the code and burn out more than I ever have.

I also trying to remove some use of the id variable like $tabel_b1_field1 and replace them with $code instead to ensure concistency.

I'm still trying to add the trigger to finally make the history table work. But I'll do it on the next commit